dqn-mountaincar/requirements.txt
gymnasium[classic-control]
torch
numpy
pandas
matplotlib
pyyaml
tqdm